Move repeated functions get_ip and get_port to shared helpers. Added pycache to gitignore.
