+++ .gitignore
@@ -1,5 +1,6 @@
 **/node_modules
 package-lock.json
+**/__pycache__
 
 # binaries
 minikube-linux-amd64
