+++ tests/test_api.py
@@ -1,16 +1,5 @@
-import subprocess
 import requests
-
-def get_ip():
-    cmd = ["minikube", "ip"]
-    return subprocess.check_output(cmd).decode('utf-8').strip('\n')
-
-def get_port():
-    cmd = ["kubectl", "get", "service", "gateway-service", '--output=jsonpath="{.spec.ports[0].nodePort}"']
-    return subprocess.check_output(cmd).decode('utf-8').strip('"')
-
-def get_url():
-    cmd = ["minikube", "service"]
+from helpers import get_ip, get_port
 
 def test_correct_use():
     try:
