+++ tests/test_subtraction.py
@@ -1,13 +1,5 @@
-import subprocess
 import requests
-
-def get_ip():
-    cmd = ["minikube", "ip"]
-    return subprocess.check_output(cmd).decode('utf-8').strip('\n')
-
-def get_port():
-    cmd = ["kubectl", "get", "service", "gateway-service", '--output=jsonpath="{.spec.ports[0].nodePort}"']
-    return subprocess.check_output(cmd).decode('utf-8').strip('"')
+from helpers import get_ip, get_port
 
 def get_url(val1, val2):
     return 'http://' + get_ip() + ':' + get_port() + f'/subtract?num1={val1}&num2={val2}'
